// app/interfaces/order.ts
export enum OrderStatus {
   OPEN = 'open',
   PENDING = 'pending',
   PAID = 'paid',
   DELIVERING = 'delivering',
   CONFIRMED = 'confirmed',
   COMPLETED = 'completed',
   CANCELLED = 'cancelled',
   REOPENED = 'reopened',
 }
 export enum DeliveryMethod  {
   PICKUP = 'pickup',
   DELIVERY = 'delivery',
 }
 
 